refactor: architecture doc fixes, test file split, stale doc cleanup
- Fix architecture.md tool count from 4 to 5 (add analyze tool)
- Split handler_unit_test.go (812→627 LOC) by extracting upgrade/update
  warning tests into handler_warning_test.go (195 LOC)
- Update known-issues.md version from v5.8.0 to v0.7.x after rebadge
- Remove stale "stub handlers" section — audit_log, diff_sessions, and
  playback_results are fully implemented with tests
